+++ src/sc/photogallery/config.py
@@ -9,6 +9,7 @@ JS_RESOURCES = (
     '++resource++collective.js.cycle2/jquery.cycle2.min.js',
     '++resource++collective.js.cycle2/jquery.cycle2.carousel.min.js',
     '++resource++collective.js.cycle2/jquery.cycle2.swipe.min.js',
+    '++resource++collective.js.cycle2/utils.min.js',
 )
 
 HAS_ZIPEXPORT = True