+++ src/sc/photogallery/logger.py
@@ -1,0 +1,7 @@
+# -*- coding:utf-8 -*-
+from sc.photogallery.config import PROJECTNAME
+
+import logging
+
+
+logger = logging.getLogger(PROJECTNAME)